\name{tcgR-package}
\alias{tcgR-package}
\alias{tcgR}
\docType{package}
\title{Useful helper functions for TCGA data.}
\description{Useful helpfer functions for TCGA data.}
\details{
\tabular{ll}{Package: \tab tcgR\cr
Type: \tab Package\cr
Title: \tab Useful helper functions for TCGA data.\cr
Version: \tab 0.1.0\cr
Date: \tab 2012-06-09\cr
Author: \tab Markus Riester\cr
Maintainer: \tab Markus Riester <markus@jimmy.harvard.edu>\cr
LazyData: \tab true\cr
Imports: \tab 
Biobase,
affy,
reshape2\cr
Depends: \tab 
R (>= 2.11.0),
Biobase\cr
License: \tab GPL-3\cr
URL: \tab https://github.com/lima1/tcgR\cr
biocViews: \tab CopyNumberVariants, Software, Visualization, SNP\cr}
}
\author{Markus Riester}

\keyword{ package }


