\name{tcgR.methylation}
\alias{tcgR.methylation}
\title{tcgR methylation}
\description{Create a copynumbR.eset input file from Level 3 methylation data}
\usage{tcgR.methylation(path = ".", file = "tcga_methylation.txt", sep = "\\t", 
    ...)}
\arguments{
  \item{path}{The path containing the Level_3 folder.}
  \item{file}{The output file}
  \item{sep}{The field separator character. See write.table().}
  \item{\dots}{Additional arguments passed to write.table().}
}



\author{Markus Riester}





